requirements.txt
python-pptx==0.6.23
dify_plugin==0.3.0